+++ ruoyi-business/src/main/java/com/diandong/domain/dto/RawMaterialDTO.java
@@ -5,13 +5,14 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
  * DTO实体类
  *
  * @author YuLiu
- * @date 2022-03-29
+ * @date 2022-03-31
  */
 @Data
 @ApiModel("DTO实体类")
@@ -82,7 +83,7 @@ public class RawMaterialDTO implements Serializable {
      * 预估进价
      */
     @ApiModelProperty(value = "预估进价")
-    private Double prePrice;
+    private BigDecimal prePrice;
 
     /**
      * 仓库id
