+++ ruoyi-business/src/main/java/com/diandong/domain/po/DishesPO.java
@@ -7,13 +7,14 @@ import lombok.Data;
 import lombok.experimental.Accessors;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
  * PO实体类
  *
  * @author YuLiu
- * @date 2022-03-29
+ * @date 2022-03-31
  */
 @TableName("wis_dishes")
 @Data
@@ -69,7 +70,7 @@ public class DishesPO implements Serializable {
      */
     @TableField(value = "dishes_price")
     @ApiModelProperty(value = "菜品价格")
-    private Double dishesPrice;
+    private BigDecimal dishesPrice;
 
     /**
      * 菜品单位
@@ -90,7 +91,7 @@ public class DishesPO implements Serializable {
      */
     @TableField(value = "pre_price")
     @ApiModelProperty(value = "预估价")
-    private Double prePrice;
+    private BigDecimal prePrice;
 
     /**
      * 产地
