+++ ruoyi-business/src/main/java/com/diandong/domain/po/OrderDetailPO.java
@@ -7,13 +7,14 @@ import lombok.Data;
 import lombok.experimental.Accessors;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
  * PO实体类
  *
  * @author YuLiu
- * @date 2022-03-30
+ * @date 2022-03-31
  */
 @TableName("wis_order_detail")
 @Data
@@ -55,7 +56,7 @@ public class OrderDetailPO implements Serializable {
      */
     @TableField(value = "dishes_price")
     @ApiModelProperty(value = "菜品价格")
-    private Double dishesPrice;
+    private BigDecimal dishesPrice;
 
     /**
      * 菜品数量
@@ -69,7 +70,7 @@ public class OrderDetailPO implements Serializable {
      */
     @TableField(value = "dishes_total_price")
     @ApiModelProperty(value = "菜品总价")
-    private Double dishesTotalPrice;
+    private BigDecimal dishesTotalPrice;
 
     /**
      * 菜品图片 图片地址（后期如果没有了怎么办）
