+++ ruoyi-business/src/main/java/com/diandong/domain/vo/DishesVO.java
@@ -7,6 +7,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.io.Serializable;
 
@@ -14,7 +15,7 @@ import java.io.Serializable;
  * VO实体类
  *
  * @author YuLiu
- * @date 2022-03-29
+ * @date 2022-03-31
  */
 @Data
 @ApiModel("VO实体类")
@@ -62,7 +63,7 @@ public class DishesVO implements Serializable {
      * 菜品价格
      */
     @ApiModelProperty(value = "菜品价格")
-    private Double dishesPrice;
+    private BigDecimal dishesPrice;
 
     /**
      * 菜品单位
@@ -80,7 +81,7 @@ public class DishesVO implements Serializable {
      * 预估价
      */
     @ApiModelProperty(value = "预估价")
-    private Double prePrice;
+    private BigDecimal prePrice;
 
     /**
      * 产地
