+++ ruoyi-business/src/main/java/com/diandong/domain/vo/OrderDetailVO.java
@@ -7,6 +7,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.io.Serializable;
 
@@ -14,7 +15,7 @@ import java.io.Serializable;
  * VO实体类
  *
  * @author YuLiu
- * @date 2022-03-30
+ * @date 2022-03-31
  */
 @Data
 @ApiModel("VO实体类")
@@ -50,7 +51,7 @@ public class OrderDetailVO implements Serializable {
      * 菜品价格
      */
     @ApiModelProperty(value = "菜品价格")
-    private Double dishesPrice;
+    private BigDecimal dishesPrice;
 
     /**
      * 菜品数量
@@ -62,7 +63,7 @@ public class OrderDetailVO implements Serializable {
      * 菜品总价
      */
     @ApiModelProperty(value = "菜品总价")
-    private Double dishesTotalPrice;
+    private BigDecimal dishesTotalPrice;
 
     /**
      * 菜品图片 图片地址（后期如果没有了怎么办）
