+++ ruoyi-business/src/main/java/com/diandong/domain/vo/RawMaterialVO.java
@@ -7,6 +7,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.io.Serializable;
 
@@ -14,7 +15,7 @@ import java.io.Serializable;
  * VO实体类
  *
  * @author YuLiu
- * @date 2022-03-29
+ * @date 2022-03-31
  */
 @Data
 @ApiModel("VO实体类")
@@ -86,7 +87,7 @@ public class RawMaterialVO implements Serializable {
      * 预估进价
      */
     @ApiModelProperty(value = "预估进价")
-    private Double prePrice;
+    private BigDecimal prePrice;
 
     /**
      * 仓库id
